requirements.txt
streamlit==1.33.0
plotly==5.22.0
pandas==2.2.2
psycopg2-binary==2.9.9
python-dotenv==1.0.1
supabase==2.4.2
gotrue==2.4.2
postgrest==0.16.7
